Rotation Transformation Matrices Work
Managed to make the 3d skeleton shape move with the sliders. I passed the value of the slider and used it in an equation with the appropriate cosine for the matrix that is used to transform the position of shapes. In this case - a rotation transformation matrix. I needed one for horizontal (left and right) and another for vertical movement